test(passport): e2e tests for 2fa
