+++ package.json
@@ -169,6 +169,8 @@
     "remark-slug": "^5.1.2",
     "rollup": "^2.33.3",
     "rollup-plugin-terser": "^7.0.2",
-    "which": "^2.0.2"
+    "which": "^2.0.2",
+    "pngjs":"^6.0.0",
+    "jsqr":"^1.4.0"
   }
 }
