+++ stacks/api/passport/authfactor/src/totp.ts
@@ -6,7 +6,8 @@ export const TotpFactorSchemaProvider: AuthFactorSchemaProvider = () => {
   const schema: FactorSchema = {
     type: "totp",
     title: "Time based One Time Password",
-    description: "Time base one time password by mobile app like Google Auth, Duo etc.",
+    description:
+      "Verification with time based one time passwords which are usually generated by apps like Google Auth, Duo etc.",
     config: {}
   };
   return Promise.resolve(schema);
